delete window prompt

+++ src/components/BlogItem.js
@@ -16,7 +16,10 @@ function BlogItem({ blog }) {
   async function deleteBlog(id) {
     try {
       const blogsDoc = doc(db, "blogs", id);
-      await deleteDoc(blogsDoc);
+      const proceed = window.confirm("Are you sure you want to delete");
+      if (proceed) {
+        await deleteDoc(blogsDoc);
+      }
       navigate("/");
       // Display a success message, navigate back, or perform other actions.
     } catch (error) {
